backend/requirements.txt
anthropic[bedrock]
opencv-python-headless
numpy<2
python-dotenv
flask
flask-cors
ultralytics
